+++ frontend/image.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react'
+import { Image } from 'react-native'
+
+const ImagesExample = () => (
+   <Image source = {require('./assets/Lighticon.png')}
+    style ={{width:80, height:80}}/>
+
+)
+export default ImagesExample
